cloud_deployment/functions/setup_vertex_ai/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.21.1
google-cloud-discoveryengine==0.11.16 